Organize MainHook.java
